window: make NautilusWindow a GtkApplicationWindow
So we can use the application menu with it.
Note that we set show_menubar always to FALSE, since we already show a
menubar ourselves.
This will probably need a little more tweaking to make sure that in
fallback mode we get the same number of options in the menus.