pathbar: remove fake root attributes
We currently allow going up from Home towards the filesystem root with a
little back arrow; in the code this is implemented by keeping track of an
additional fake root.
Since we want to change the pathbar not to show the back arrow anymore
in such cases, remove this code (and refactor other pieces of code
around it).

https://bugzilla.gnome.org/show_bug.cgi?id=619616