view: add a shadow to the view when not showing tabs
When we are showing tabs, the notebook will take care of drawing a
border around the view.
When there is only one view, since we disable tabs from the notebook,
the view needs to draw a shadow instead, or we won't draw any shadow at
all. In the past this wasn't a problem because the toolbar used to draw
a stroke for us.