window: simplify toolbar actions creation
Instead of having the toolbar create a different action group, with
different actions, and having to keep those in sync with the window's
main action group manually, just pass the main action group to the
toolbar.
This also makes NautilusNavigationState useless (and removes a lot of
boilerplate code), and fixes Ctrl+1/Ctrl+2 not working correctly as
shorcuts to switch view type.