Don't verify symlinked target for I/O operations
Nautilus verifies the destination for free space, read-only ability
and such. This however doesn't work for symlink pointing to a directory.

Internally (in gio) we call statfs() but that actually stats the fs the
symlink is on, instead of the place symlink is pointing to. This appears
to be a limitation of POSIX statfs().

To work around this limitation, let's disable these extra checks when
destination is a folder and symlinked.

The bug manifestation are false error messages preventing to create a file
or directory when parent fs is read-only but contains a symlink to rw fs
which is the desired destination.

https://bugzilla.redhat.com/show_bug.cgi?id=844713