application: don't open the default display when parsing options
Opening the default display when parsing options means we'll do it in
non-primary instances of the application.
When the default display is opened, _gdk_x11_display_make_default() will
be called, which reads and removed DESKTOP_STARTUP_ID from the
environment - but we need to keep DESKTOP_STARTUP_ID around until
add_platform_data() is called in the non-primary instance, which takes
care of forwarding it to the primary instance.

This fixes the busy spinning cursor being stuck when opening multiple
Nautilus windows.

https://bugzilla.gnome.org/show_bug.cgi?id=664147