application: display an error dialog when connecting to server fails
Since we always return TRUE from the nautilus_window_go_to_full()
callback, we fail to show the error dialog in case connecting fails, and
we print a warning to command line.
In case there's an error, remove the warning and return FALSE instead,
to report the error to the user.