search-engine: wait for all engine's completion before starting
When we're getting reloaded, we should wait for the completion of all
engines before triggering a new search. When this happens, set an
internal restart flag and apply it when the last provider returns.
This avoids finished signals from providers that complete the previous
search after a while to being interpreted for the current search.