slot: make sure to walk the hierarchy up only for native and non-mounts
When we detect a deletion, we try to find an existing parent in the
hierarchy to display in the window.
While this is desirable when a file is deleted, we never want to walk a
non-native hierarchy, or to walk up a hierarchy from a mount root
being removed.

This fixes the window redirecting to /run/media/$username/ instead of
$HOME when a mount is removed.