slot: use the default mount location when switching to a mount root
Add a new flag to signal that we really want the default mount location
to be loaded if possible. This way we can still get to the mount root
e.g. by going up or typing it directly in the location bar, but the
default location between sidebar mounts and using Connect to Server will
be the same.

https://bugzilla.gnome.org/show_bug.cgi?id=522917