doc folder cleanup.
A lot of the files in the doc/ folder are no longer relevant as
most of them are over ten years old.

architecture.txt

Some details that include metaviews and CORBA.
This is completely outdated.

nemo-internals
Has a lot of information about the inner workings of
nautilus. However, it was written in 2000. Also severely outdated and
unnecessary.

nemo.faq

Only contains a short introductory text on what nautilus
is/was.

recommended-books.html

contains recommended books on c++ and gtk+
development. Current sources have no c++ code and the gtk+ book
referenced in the file was written in 1999. Not only irrelevant
to the current nemo project but also completely out of place.

smoketests.html

"Nemo Smoke Tests are a minimal set of tests, to determine whether
a build is usable for testing, and to catch obvious regressions."
Last changed in the year 2000. Also outdated and irrelevant.

state-machines.txt
Proposal for the loading state machines.
ViewFrame state machine.

There's no mention of ViewFrame in the current sources.