Couple tweaks to speed up item selection a bit:
Increased menu recalculation timeout a bit.  Could probably go more,
since the menu gets force-recalculated any time the context menu is
activated, but it's out of the way at least, and shouldn't interfere
with selection and navigation speed.

Also changed statusbar update from an idle timeout which was basically
instantaneous to the selection changing, involving a very elaborate
string manipulation routine, to a 100ms timeout.  Prevented updating
if the selection is too fast (i.e. holding arrow down) which is useless
anyhow, but it updates immediately upon stopping on an item.  The status
blanks out during a repeat so as not to be confusing.