Switch to next tab after closing the active one
Closing the active tab would previously select the first inactive slot
in a pane. It is more common in other applications though to select
either the next slot (if one exists) or the previous one. Note that we
have to consult the GtkNotebook which holds the slots for this instead
of the pane's list of slots since the latter does not necessarily
represent the order in which slots are actually displayed in Nemo.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>