NautilusPropertyDialog: Fix user list for owner change
This code was constantly failing to find the current files user as
the form of the "owner" attribute didn't match the internal list.
Also, as a consequence of this the user list didn't match the old
one so we keep re-creating it all the time.

We fix this by being much more explicit in what we store in the
model columns:
0 - the display name
1 - the return value from nautilus_get_user_names()
2 - the actual username (not real name)