nautilus-list-view: Avoid unreadable names
Currently if the view is resized, the column name is resized as well
given that use ellipsization allowing the column to become unreadable.
To avoid that, use width-chars property to set a desired width, but at
the same time allowing the user to resize without limits the name column
if desired.

https://bugzilla.gnome.org/show_bug.cgi?id=732004