Don't register bookmarks_breakpoint as gproperty
The bookmarks_breakpoint was previously registered as gproperty, but not
consistently treated as such (e.g., by directly assigning to it rather
than calling `g_object_set()`). It's enough to listen for modifications
of the respective gsettings key; we don't need to register the property
with the NemoPlacesSidebar class as well.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>