nautilus-application: Parse force-desktop before exiting
In commit ae4d4960 we introduced a regression that a new window was
openned if both --no-default-window and --force-desktop options were
used.

To avoid that, activate those options before actually skipping the
activate of the application if --no-default-window option is provided.

The application nornally would exit if --no-default-window is provided
and the show-desktop-window is not activated, but, we rely on a the
detail that activating the open-desktop action when --force-desktop is
provided as a option, creates a new window, which makes the application
keep alive.

https://bugzilla.gnome.org/show_bug.cgi?id=741166