GTK 3.20-fix wrong desktop size
Reset default size of desktop window to -1 in GTK 3.19/20 only. The combination of a non-resizable
window with a default size is what causes the desktop to open wrong size.

based on GNOME/nautilus@cf79068

"desktop-window: avoid wrong size"

"gtk+ recently changed and unresizable windows now take into account
the default size set. We set the default size for normal resizable
windows in nautilus."

"However the desktop window is a no resizable window. What happens now
when setting a default size is that that size is used instead of the
widget request size, since window sizing now takes precedence over
widget sizing."

"To fix it, set the default size to -1 to override what the nautilus
window does by default, so we can avoid this misbehaviour."