Fix --geometry option when Nemo is already running
Fix for https://github.com/linuxmint/nemo/issues/569 based on 
https://github.com/mate-desktop/caja/pull/618/commits/9fb12116c1a09221a88f5c1a7e9010022c96df66

This fixes a GApplication command line issue inherited from Nautilus 3.4. The issue is that using --geometry in a second invocation of nemo sets  self->priv->geometry in the local instance, but the windows are opened by the remote (already running) instance. That reads it's own version of self->priv->geometry, which is not written to by setting the same variable in the local instance. There are very few ways to pass data from a local command line to "open" but Gapplication provides the "hint" string variable. It can pass the geometry string, which can then be written to  self->priv->geometry, making the geometry option work again.