nemo-application.c: fix maximized window state preservation
This fixes the maximized window state not being preserved properly since
PR #1261.

While trying to track this down, I found that we are in some cases passing
NULL to g_application_open in the hint parameter which is explicitly
forbidden and results in unexpected strings being passed to
nemo_application_open. I have added a NULL check there to prevent this.

Aside from that, I also had to add a check for an empty string in open_window
otherwise we override the previously restored window state in too many cases.