nemo-file.c: add EFAULT as a valid exception to the access_ok function.  This can throw when the file has just been moved or is in the midst of being moved and the internal async updates have not completed yet (usually during a fairly large transfer.)  As a result, the icon cache problem bar appears needlessly.
The original purpose for the thumbnail cache detection is that programs
that are run with elevated privileges (such as editors) will generate
their own thumbnails with elevated priviliges, in the common folder
designated for such by both gnome-desktop and cinnamon-desktop, which
both provide the thumbnailing API that many programs use.

Unfortunately, when nemo attempts to update the thumbnail for a
particular file, and cinnamon-desktop reports that it cannot, because
of invalid permissions on an existing cached thumbnail, there is really
nothing that can be done except either a) regenerate the entire cache or
b) repair the broken files.  nemo --fix-cache repairs any broken files,
clearing the issue, and allowing thumbnailing to proceed on the file.

Fixes: #1429
Fixes: #1248
Fixes: https://bugs.launchpad.net/linuxmint/+bug/1592293