nemo-action.c: Use non-blocking checks for a directory.
The call in the get_visibility routine could and would
cause ui freezes when working with remote servers, and the
connection was lost.

I believe the original reason for these calls (which replaced
nemo_file_is_directory) was that a NemoFile may not always have
the most up-to-date information on a file - it relies on file
listeners and and updates metadata lazily.  We need to know
whether something is a directory sooner, such as when setting
up our sidebar view.  And in the case of remote shares, we
don't load *any* metadata unless the connection is there (even
though this metadata is stored locally, regardless of the file's
location.  This is a gvfs limitation)