Rename show-superuser-toolbar into show-root-warning
And remove the UI frontend for that preferences.

Distros can glib override it and advanced users can set it via gsettings, but
it shouldn't be available in the UI prefs, because it doesn't work as expected.

The settings are user-specific thus interacting with the UI prefs has no impact
on the app running in elevated mode. Depending on whether nemo was started with
sudo, pkexec, gksu or gksudo, different behaviours can be expected.