[gtk3.22] Pathbar style (#1765)
*  nemo-pathbar: Better support for gtk3.22

Add new "path-bar" and "slider-button" style classes. This simplifies theming
since it allows the path bar to pick up the styles that already exist in most
themes for the GtkPathBar.

While we're at it stop using deprecated GtkArrow in the slider buttons and use
a standard image instead.

* nemo-location-bar: Removing a bit of the hard-coded padding

This forces all the toolbars to minimum height. To keep square buttons in the
toolbar then requires a good bit of horizontal padding and leaves us using much
larger buttons than we really need.