Revert "also use exec-arg from org.cinnamon.desktop.default-applications.terminal when using "open in terminal""
This reverts commit 7ca7990c0c33fa4a4f19ba0f534f00f221acb480.

It was a mistake to use the args key in Nemo. This is used by Cinnamon when
running something in a terminal (and there, -x is needed).

It shouldn't be used here in Nemo. In the context of Nemo we're only interested
in opening a terminal, not running something in it.