nemo-view.c: Don't try to update undo menu items in a non-active slot.
When a slot becomes un-focused, the menu manager clears all the action
groups.  When the slot regains focus, this is reversed, and the menus
are updated.  The undo manager is a singleton that sends updates to
all listeners (all views.)  Without blocking and unblocking during
the menu unmerge/merge cycle, the simplest thing to do is check for
the required action group being NULL during the undo manager callback,
and ignore the callback it is.