trash:  Show the correct icon in the pathbar, and fix activation uri for folders located in the trash.
We don't need a trash monitor in the pathbar, the icon updates thru
other means, and gets the correct icon.

When opening a folder inside the trash location, it would revert to
displaying the file as e.g. /home/user/.local/share/Trash/files/file
instead of trash:///file - this would remove the trash bar and
mess up the pathbar display (showing the absolute path instead of
relative to trash:///.)