eel-gtk-extensions.c: Fix menu popups when a valid event is not provided.
gtk_menu_popup_at_pointer fails when triggering a popup using the
keyboard menu key or ctrl-F10. Go back to using the old, deprecated
function in these cases where a valid event is not provided.

According to the documentation, the event can be null, in which case
the current event will be used.  However, it fails to get a valid
GdkWindow from this event.

Fixes #1897