nemo-statusbar.c: Remove custom scroll handler.  The gtk one works correctly now.
Long ago, in 37adad145e5, we added a custom scroll handler because the
Gtk one behaved the reverse of what was logical at the time.  This
appears to have been fixed in Gtk, so we don't need this handler. Tthis
will fix the annoying tendency of nemo always zooming out first, before
zooming in, as the scroll delta of the mouse wheel resets each time focus
is lost on the scroller.  It seems that the initial scroll is ignored
by gtk (probably to address this very problem) until a delta actually
exists (on scroll events after the first one.)