menus: Respect the gtk-menu-images preference for the primary 'open/ open-with' menu item.
If gtk-menu-images is false, this menu item is usually the only one
visible in the selection context menu.  If the toggle switch at the
right end of the item was hovered, part of the display effect was to
hide the menu item image temporarily.  When this occurred, Gtk noticed
there were now no images *at all* in the menu, and would decide to
resize, causing a jitter when entering and leaving the toggle icon.

This commit stops forcing the menu item to always show an image,
which has a side effect of fixing the issue described above. Otherwise,
we'd need to supply a fake 'blank' icon to maintain the menu size
when the real image was hidden.

Other changes fix the gtk preference only having an effect on the
running instance when turned on, but not when turned off.

Fixes #1370