nemo-mime-actions.c: Don't offer the 'Make executable and run' option by default.
There's really not much sense in this.  Any unknown file will nearly
never be some innocuous script (Gio's mime sniffer would simply
classify such as a text file.)  The only sensible option here is to
choose a program to open the file.

In the event this actually breaks someone's 'flow' there's a hidden
setting to re-enable it.

Fixes #1923