nemo-mime-application-chooser.c: Remove hacks for file-picker button.
A GtkFileChooserButton wants to display an icon and label corresponding
to the selected file.  We were hiding these unwanted items, since we
end up putting the selected program in the custom entry box instead.
In the latest Gtk, there's now a separator in the button, so it would
be another widget to hide.

This switches to using a simple button with a separate file-picker
dialog that we generate, and allows us to remove all hacks.