nemo-application.c: Add minimally necessary theme elements to the current theme if they appear to be missing.
Things such as the shaded inactive pane style when using dual-pane, and
rename editable styling for the icon view.  Not having these, regardless
of the remaining nemo style elements, can cause a great deal of confusion,
not to mention spurious bug reports.

Since this operation has a slight performance impact on startup, we may as
well only check themes that we're not already confident of having nemo
support.