nemo-list-model.c: Check for unreadable folders and remove the expander when one is discovered during the async count job.
A folder inaccessible to the user should not allow toggling the
expander.  We need to check for this unreadability during the initial
loading of items, since some other window or view may already have
this folder's contents loaded, allowing the new view to display
accurate count info right away.

This fixes a regression caused by the following commit, which would
cause a segfault when attempting to expand into an inaccessible folder.

https://github.com/linuxmint/nemo/commit/9d80ffd8eb048d5636b7de24325f6598380dd8d5