nemo-icon-container.c: Don't recalculate the layout while renaming a file.
When a file enters rename mode, its label can change size slightly,
temporarily, which can trigger a size_allocate call.  When the size
is near the height boundary that will remove the scrollbars, size
allocation calls end up sent in pairs, the first reducing the width
allocation, and the second restoring it.

The width change causes the icon layout to change, which ends rename
mode immediately.  This is observed as seeming as though rename mode
never occurred in the first place.

This will ignore the first allocation in rename mode, and only do a
relayout on subsequent allocations, so we skip the resize/termination
of rename, but behave normally for any further allocations (such as
the user actually resizing the window.)

This issue cannot be reproduced when using gtk overlay scrollbars, as
they have no impact on allocations.

Fixes #1948