nemo-file-management-properties.c: Don't drop builder object ref until the window is closed, as callbacks to setting changes refer to it still.
Related, the change handlers for the Icon Captions preferences were
out of order in nemo-icon-view.c and nemo-icon-view-container.c.
The view was refreshing before the new setting was fetched by the
container object.

We add the setup to the container _init function instead of getting
it when text is requested while adding icons to the view.  Both
the view and containers connect to the same signal, for different
reasons, but now the container connects first (and will also get
called first, so the setting is up-to-date when the icon view calls
for a refresh.

Fixes #1983