places sidebar: Enable actions to work properly even when the item being right-clicked/context-menu'd is not the currently loaded view.
Previously, for things such as removable drives, one had to select
(left-click) the item before any drive-related actions (like format)
would appear.

This is because there is no underlying NemoFile for this location
prior to calculating action visibility, and as it's created on demand,
its enclosing mount is not known, causing the action to be unavailable.

Once the location is selected, the view is switched to that location,
and the file object is fully populated (including mount info).  At
this point, the action would work.

With this change, the action visibility check will find the enclosing
mount for local files, and set it in the NemoFile object, for later
use upon activation of the action.  A ref is kept on the object so
that the file isn't destroyed right after the visiblity check and
recreated for the action activation (which would be a new object,
without the mount info).

Note: g_file_find_enclosing_mount() is potentially blocking, so this
is only applied in sidebar cases, and only on local files.