nemo-window.c: Fix hidden menu alt-key activation.
The existing behavior for alt-activation of the menu was interfering
with other shortcuts, including window/workspace navigation, and
generally was counterproductive.

New behavior has the menu showing on alt release, and hiding on press.
Additionally, the showing will grab the focus so a mnemonic can be
typed to activate one of the submenus.  Extraneous keystrokes between
a single alt press and release will cause a pending menu show to be
skipped, so we don't get the menu appearing when an unrelated shortcut,
that involves the alt key as a modifier, is activated.

This behavior brings it more in line with other applications that offer
hidden menubar behavior (such as firefox).

Fixes #1389
Fixes #1895