nemo-places-sidebar.c: Keep the selected bookmark scrolled into view when a selection is made on a bookmark.
Fixes (partially) #1968

Ideally, the sidebar position doesn't move (as is currently the case
when selecting a mount or other non-bookmark item).

The problem is that we get a lot of change events from our bookmark
list when a bookmark is selected (from file info changing, etc..) -
much of this can be pruned out, and the updates reduced in a future
patch, since we no longer allow custom icons in the sidebar, the only
real reason for change is for remote bookmarks when they become online
or 'available' and their icon changes.

This commit at least ensures the selected item is kept centered in the
view (to the extent possible, depending on where it is in the list).