nemo-list-view.c: Don't use fixed-height mode, it does not work well with zoom changes.
This reverts commit 1b068a1dc6341aca236525136231c20e66fb959a.

It had appeared there were no ill effects when zooming with fixed
mode on, but when navigating a folder with a large number of files,
only the files visible in the current scroll region had their
heights and spacing adjusted.  Out-of-view items would become corrupt
and break the view entirely as you scroll up or down from the original
position.

Attempting to enable the mode only during loading of a folder worked,
until the mode was disabled, when any performance improvement that had
been gained would be lost as the view would update itself.  So it's
better off just removing it.

Fixes linuxmint/mint-19.1-beta#1