nemo-list-view.c: Always display at the default list view zoom level during a search.
When ignore-metadata is active, the list view was re-using the previous
view's zoom level.  This makes the search view consistent regardless of
the state of ignore-metadata.

It might be desirable to add a search zoom level preference (similarly
to how we currently store the search columns).

Fixes linuxmint/mint-19.1-beta#55