nemo-view.c: Don't ignore non-symbolic extension menu icons.
When symbolic icon display was fixed in fc1194a8ea0a6b8586af8e we
also chose to ignore any icon names that weren't symbolic.  Since
we've updated all our extensions to use symbolics anyhow, there's
no real reason to enforce this on third-party extensions (since we
never enforced it for actions, anyhow).

Restore the ability to use any type of icon, as well as the ability
to use an absolute icon path instead of name (which was also lost
in that commit).

Fixes #2001