nemo-icon-info: Use different text widths for the desktop than for windowed mode.
This fixes extremely narrow text fields on the desktop for larger
icon sizes.

This value should eventually be dynamic to a certain extent,
depending on the spacing between items on the desktop, but there
needs to be some significant refactoring to make all of the required
values accessible.  For now this provides some improvement over the
previous state.