nemo-action.c: Allow discovery of tokens in any order in an action line.
As of d41298bc747df3de, the find_token_type() function was only
finding tokens in the order listed in the function.

For example, if an action exec line was 'foo %P %F', since the parent
token (P) came after the file path token (F) in find_token_type(),
it was completely ignored.

This changes the function to search for occurrences of '%', then
performing a prefix check starting at the result pointer.

Should fix issue one in #2107.