nemo-window-manage-views.c: add a weak pointer to the window slot's bad cache bar when spawned.
We use the cache_bar reference to keep track of whether or not we've
displayed it in the current slot, but its lifetime is managed thru
add/remove_extra_location_widget() and there's nothing to null out
this member when the widget is removed (infobars are discarded when
the location changes currently).

This will set it to NULL when the bar is finalized (which happens
when it's removed from its parent).

Fixes an un-dismissed cache bar not re-appearing after the first
time it's shown.