Navigation: Fix go to parent action for Samba
g_file_get_parent() only works within a Samba share.
When asking for the parent of smb:///, smb:///workgroup,
smb://hostname, smb://hostname/sharename, we get NULL.

This commits addresses this limitation so that we can return
to network:/// from smb:///, return from smb:/// from the workgroup
or a host, and return to the host from a share.