nemo-mime-application-chooser.c: escape file names in Open With tab
The creation of the label "Select an application in the list to open [...]" in the "Open With" file properties tab fails because of file names which include HTML special chars like ampersand or angle brackets.
I wasn't able to inject HTML into the label because of forbidden slashes in file names. HTML comments (<!-- -->" do work though!

Replaced g_strdup_printf with g_markup_printf_escaped to fix that.