nemo-*-view.c: Update the selection info when a view has finished loading.
When a view was re-used (this happens when navigating from one folder
to another, where both folders are set to icon view), the selection
from the previous location was being kept until a new selection was
made.  This caused action visiblity calculations to be incorrect if
it was a 'background' click action.

Fixes #2167