list view: Don't remove the dummy (Loading) entry except during post-loading updates.
When expanding a directory row, the directory is re-checked, to
determine whether there are actually any children to add to the
model. This calls the model's file-changed handler, which could
remove the existing dummy row in the process, causing the model
to become corrupted.

Avoid removing this dummy row while expanding the parent, regardless
of whether or not the parent has any actual children.  If it turns
out there are no longer children to load, show the traditional
"Empty" row.

Fixes linuxmint/nemo#2256

Note: This prevents the crash, but does not fix the behavior of
the parent folder not knowing it has gained or lost it's only child.
This is a limitation due to the fact that the currently visible
tree items are only monitored as 'files', not folders - folders get
an initial count when loading the view, allowing the expander to
be added (or not).  This is pre-existing behavior.  It may or may
not be possible to change this in the future.