nemo-window.c: Monitor and react to scale-factor changes.
The desktop was adapting properly because it had a listener for
scale-factor already, but regular nemo windows didn't, and would
not resize items in the icon view (since that is a custom view,
not a proper GtkIconView).

Adding this to nemo-window solves the issue for both the desktop
as well as normal windows since they all inherit from NemoWindow.