icon view: Process thumbnails/previews only after view is fully loaded.
Fix new window layout - the new window was not calling for the layout
calculation because the value wasn't initialized to -1.

Implement list view thumbnail delay.

nemo-list-view.c: Reverse the order of visible row reveals, so they're
processed top-to-bottom.

nemo-icon-container.c: clear queue pointer so it will get remade
when the new location loads.