thumbnails: generate/load thumbnails on visible items only, and drop background loading.
Views load much more quickly when delaying any thumbnail processing
until all items are fully loaded.

An attempt is made to load 'ahead' and 'behind' the visible range,
but in the list view, the current api does not allow scanning for
rows 'above' the current view position, only below.  As a result,
scrolling a view down in the list view will reveal icons that have
already been queued to load their thumbnail, but scrolling up will
not.  This behavior works properly in the icon views in both
directions.