data/nemo.desktop.in: fix icon
Replace icon "folder" with "system-file-manager". Issue #1686 discussed
already, that it should not use "folder" to follow free desktop
standards, and it was changed to "nemo" in 5d4ca0 ("Rename Nemo icon
from places to nemo"). It looks like it was changed back by accident in
099421 ("l10n: Update POT file and generate files").

Change it to "system-file-manager", because that makes it pick up a nice
icon with the default GNOME theme. Adjust "generate_additional_file"
too, so it won't get overwritten by a accident again.