nemo-main-application.c: Remove 30s inactivity timeout.
This was originally in place to prevent the nemo process
from constantly being restarted if activated and used
over dbus.

The nemo-desktop now runs in the background and claims
ownership of the org.freedesktop.FileManager1 service, and
since it autostarts it will prevent nemo from doing so.

In the event someone has disabled nemo-desktop entirely,
nemo still 'advertises' as the dbus service so would launch
to fill this need in that situation. This is an uncommon
situation so it's probably ok to get rid of the inactivity
timeout.