nemo-view.c: Only update the context menu when it's about to be shown.
Previously, the menu was updated when the selection changed, during
loading of the view, when a file in the view (or the view location
itself) changed and many other events. This would end up causing
a lot of unnecessary updates, which could have a noticeable impact
on performance.

Instead, only flag the menu as needing an update when any of these
events occurs, and only actually update the menu just before showing
it.

ref: #2718