nemo-view.c: Delay actions and extension menu providers from updating until the menu is requested.
This is a compromised version of 6269be562820 (which is now
reverted) that will still allow built-in menu items (and their
associated keyboard shortcuts) to be managed when the selection
changes, but the potentially expensive checks that extensions
and actions can make are delayed until the user requests the
menu.

This seems counter-intuitive, but update_menus occurs fairly
frequently for other reasons besides the selection changing,
and the costlier updates can impact the ui significantly.

On the other hand, the built-in menu items are fairly quick,
so updating them frequently isn't as big a deal, and is needed
for basic things like delete and rename shortcuts.