nemo-directory.c: Fix memory leak.
When creating a new file or a new file from a template, it's possible
that file->details->is_added may not be TRUE, and cause this file to
not be finalized along with other files if the view directory is
destroyed.

This can cause issues when re-entering that directory, with the file
being in an undefined state, and could prevent the view from fully
loading the location.

The symptoms are similar to #2623

Fixes #2736