l10n: Remove msgids
In English this looks OK and more or less consistent.

In French and probably other Latin languages this causes a problem because the subject of the imperative isn't the same:

- "Enter text" --> this is addressed to the user.
- "Search in "%s" --> this describes what the program is about to do, if anything it's imperative addressed to the computer/program.

Typically in French we would use the 2nd person to conjugate the first sentence in the imperative tense:

"Entrez.."

but we'd use the infinitive to describe the action or what we want the program to achieve:

"Chercher"

This makes the two entries very inconsistent and creates confusion.