nemo-mime-actions.c: Only add entire view file list to an image viewer launch if its commandline supports multiple files.
As a result of https://github.com/linuxmint/nemo/pull/2705, when
a single image file is selected and launched, the full file list
for the current directory is appended for certain viewers.

This is a problem if the viewer's desktop file  Exec line doesn't
support multiple files, as it can cause multiple instances of
a program to launch when that was not the intent.

Only apply this new behavior if the Exec line accepts multiple
files (%U and %F).

ref: https://github.com/linuxmint/nemo/issues/2771