nemo-file.c: Remove a file's 'failed thumbnail' marker when refreshing, if it has one.
When a thumbnailer fails on a particular file, it creates an entry
in ~/.cache/thumbnails/failed/ which prevents any future attempts
to thumbnail that file (unless it gets modified).

When the user 'refreshes' the folder, one thing it does is remove
any existing file thumbnails so they'll be regenerated. Any failed
marker should also be removed so all files can be regenerated, not
just previously successful ones.