nemo-window-manage-views.c: Only close a window when its location is deleted if its immediate parent was the desktop and the window also has no backward history.
With b3e5e170ca8, if a window's current location was deleted, the
window would close if it had no backward history. The problem is
it's not clear why this would happen - if this happens on a window
that has been used for navigation, the view will simply change to
the parent instead of closing.

Instead, target desktop folders only, which is what the original
commit was mostly to address.

This has the potential to still occur in some other situation,
mostly involving direct navigation to the folder to be deleted,
rather than opening from the desktop itself, but it's pretty
small.

Fixes #2912