interface: Add copy as path item
This simple feature adds a MenuItem to the MenuBar and Right-Click
context menu that copies the filepath to the clipboard. This is only
available if anything is selected. On local files, it takes the
regular path and on remote files (e.g. smb or ftp) it takes the uri.
If multiple files are selected, all paths/uris are pushed together
into one string with a comma as seperator.